colblk: widen Decode function offset/count types to prevent overflow
The Decode functions in the colblk package (DecodeBitmap, DecodeRawBytes,
DecodePrefixBytes, DecodeUnsafeUints, DecodeUnsafeOffsets) previously used
uint32 for offsets and int for counts. Arithmetic on these values (e.g.
off + uint32(rows*w)) was prone to overflow.

Change the offset parameter and return type from uint32 to uint64, and the
count/rows parameter from int to uint32. This eliminates overflow in offset
arithmetic while keeping counts bounded to the uint32 range which is
sufficient for row counts. Note that offsets are checked against the
length of the buffer so they can't get large enough to cause 64-bit
overflows.

Also add bounds checks: an overflow check for count+nBundles in
DecodePrefixBytes, and an endOffset vs len(b) check in DecodeRawBytes.

The DecodeFunc type, DecodeColumn, and pageStart are updated accordingly.
All callers are updated with appropriate casts.

+++ cockroachkvs/cockroachkvs_test.go
@@ -211,10 +211,10 @@ func TestKeySchema_KeyWriter(t *testing.T) {
 			for i := 0; i < kw.NumColumns(); i++ {
 				offs[i+1] = kw.Finish(i, row, offs[i], b)
 			}
-			roachKeys, _ := colblk.DecodePrefixBytes(b, offs[cockroachColRoachKey], row)
-			mvccWallTimes, _ := colblk.DecodeUnsafeUints(b, offs[cockroachColMVCCWallTime], row)
-			mvccLogicalTimes, _ := colblk.DecodeUnsafeUints(b, offs[cockroachColMVCCLogical], row)
-			untypedVersions, _ := colblk.DecodeRawBytes(b, offs[cockroachColUntypedVersion], row)
+			roachKeys, _ := colblk.DecodePrefixBytes(b, uint64(offs[cockroachColRoachKey]), uint32(row))
+			mvccWallTimes, _ := colblk.DecodeUnsafeUints(b, uint64(offs[cockroachColMVCCWallTime]), uint32(row))
+			mvccLogicalTimes, _ := colblk.DecodeUnsafeUints(b, uint64(offs[cockroachColMVCCLogical]), uint32(row))
+			untypedVersions, _ := colblk.DecodeRawBytes(b, uint64(offs[cockroachColUntypedVersion]), uint32(row))
 			tbl := tablewriter.NewWriter(&buf)
 			tbl.SetHeader([]string{"Key", "Wall", "Logical", "Untyped"})
 			for i := 0; i < row; i++ {
